requirements.txt
Django==2.2.4
django-cors-headers==3.0.2
django-jsonfield==1.3.1
djangorestframework==3.10.2
pkg-resources==0.0.0
pytz==2019.2
sqlparse==0.3.0
